gitlab-handbook-chatbot/requirements.txt
streamlit==1.38.0
trafilatura==1.8.0
beautifulsoup4==4.12.3
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
google-generativeai==0.7.2
pydantic==2.8.2
tqdm==4.66.4
